Projekat 1 i 2/notification/requirements.txt
grpcio==1.46.3
grpcio-tools==1.46.3
protobuf==3.20.1
six==1.16.0